+++ service/cloudwatchlogs/errors.go
@@ -25,7 +25,8 @@ const (
 	// ErrCodeInvalidSequenceTokenException for service response error code
 	// "InvalidSequenceTokenException".
 	//
-	// The sequence token is not valid.
+	// The sequence token is not valid. You can get the correct sequence token in
+	// the expectedSequenceToken field in the InvalidSequenceTokenException message.
 	ErrCodeInvalidSequenceTokenException = "InvalidSequenceTokenException"
 
 	// ErrCodeLimitExceededException for service response error code