+++ service/sts/examples_test.go
@@ -79,6 +79,48 @@ func ExampleSTS_AssumeRole_shared00() {
 	fmt.Println(result)
 }
 
+// To assume a role using a SAML assertion
+//
+
+func ExampleSTS_AssumeRoleWithSAML_shared00() {
+	svc := sts.New(session.New())
+	input := &sts.AssumeRoleWithSAMLInput{
+		DurationSeconds: aws.Int64(3600),
+		PrincipalArn:    aws.String("arn:aws:iam::123456789012:saml-provider/SAML-test"),
+		RoleArn:         aws.String("arn:aws:iam::123456789012:role/TestSaml"),
+		SAMLAssertion:   aws.String("VERYLONGENCODEDASSERTIONEXAMPLExzYW1sOkF1ZGllbmNlPmJsYW5rPC9zYW1sOkF1ZGllbmNlPjwvc2FtbDpBdWRpZW5jZVJlc3RyaWN0aW9uPjwvc2FtbDpDb25kaXRpb25zPjxzYW1sOlN1YmplY3Q+PHNhbWw6TmFtZUlEIEZvcm1hdD0idXJuOm9hc2lzOm5hbWVzOnRjOlNBTUw6Mi4wOm5hbWVpZC1mb3JtYXQ6dHJhbnNpZW50Ij5TYW1sRXhhbXBsZTwvc2FtbDpOYW1lSUQ+PHNhbWw6U3ViamVjdENvbmZpcm1hdGlvbiBNZXRob2Q9InVybjpvYXNpczpuYW1lczp0YzpTQU1MOjIuMDpjbTpiZWFyZXIiPjxzYW1sOlN1YmplY3RDb25maXJtYXRpb25EYXRhIE5vdE9uT3JBZnRlcj0iMjAxOS0xMS0wMVQyMDoyNTowNS4xNDVaIiBSZWNpcGllbnQ9Imh0dHBzOi8vc2lnbmluLmF3cy5hbWF6b24uY29tL3NhbWwiLz48L3NhbWw6U3ViamVjdENvbmZpcm1hdGlvbj48L3NhbWw6U3ViamVjdD48c2FtbDpBdXRoblN0YXRlbWVudCBBdXRoPD94bWwgdmpSZXNwb25zZT4="),
+	}
+
+	result, err := svc.AssumeRoleWithSAML(input)
+	if err != nil {
+		if aerr, ok := err.(awserr.Error); ok {
+			switch aerr.Code() {
+			case sts.ErrCodeMalformedPolicyDocumentException:
+				fmt.Println(sts.ErrCodeMalformedPolicyDocumentException, aerr.Error())
+			case sts.ErrCodePackedPolicyTooLargeException:
+				fmt.Println(sts.ErrCodePackedPolicyTooLargeException, aerr.Error())
+			case sts.ErrCodeIDPRejectedClaimException:
+				fmt.Println(sts.ErrCodeIDPRejectedClaimException, aerr.Error())
+			case sts.ErrCodeInvalidIdentityTokenException:
+				fmt.Println(sts.ErrCodeInvalidIdentityTokenException, aerr.Error())
+			case sts.ErrCodeExpiredTokenException:
+				fmt.Println(sts.ErrCodeExpiredTokenException, aerr.Error())
+			case sts.ErrCodeRegionDisabledException:
+				fmt.Println(sts.ErrCodeRegionDisabledException, aerr.Error())
+			default:
+				fmt.Println(aerr.Error())
+			}
+		} else {
+			// Print the error, cast err to awserr.Error to get the Code and
+			// Message from an error.
+			fmt.Println(err.Error())
+		}
+		return
+	}
+
+	fmt.Println(result)
+}
+
 // To assume a role as an OpenID Connect-federated user
 //